I added support for binutils 2.20. I added a command line argument for the gui program letting the user set a vesa mode by its number. I added boot.sh that runs on boot. Currently it runs the command "gui 11A" which switches SollerOS to 1280x1024x16. I fixed a few memory corruption errors caused by stack overflows.

+++ trunk/cross/newlib.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
-#if [[ $UID -ne 0 ]]; then
-#	echo "$0 must be run as root"
-#	exit 1
-#fi
+if [ $UID -ne 0 ] && [ $UID -ne 1000 ]; then
+	echo "$0 must be run as root"
+	exit 1
+fi
 svndir=$PWD
 cd /SollerOS/src || exit 0
 rm -rf build-gcc-full build-newlib || exit 0
@@ -16,16 +16,16 @@ export PREFIX=/SollerOS/cross
 mkdir build-newlib build-gcc-full
 cp -r --remove-destination $svndir/*/ . || exit 0
 
-#echo Rebuilding autoconf caches
-#cd newlib-1.17.0/newlib/libc/sys
-#autoconf || exit 0
-#cd solleros
-#autoreconf || exit 0
+echo Rebuilding autoconf caches
+cd newlib-1.17.0/newlib/libc/sys
+autoconf || exit 0
+cd solleros
+autoreconf || exit 0
 
 export PATH=$PATH:$PREFIX/bin
 
 echo Building Newlib
-#cd ../../../../../build-newlib
+cd ../../../../../build-newlib
 cd build-newlib
 ../newlib-1.17.0/configure --target=$TARGET --prefix=$PREFIX || exit 0
 make || exit 0
